Allow comma separated list of priceIds
