+++ .env.example
@@ -1,6 +1,7 @@
 APPLE_TEAM_ID=ABCDE12345
 NEXT_PUBLIC_BUNDLE_IDENTIFIER=com.example.yourapp
 
+# Either a single price_id or a comma separated list of multiple ids
 NEXT_PUBLIC_PADDLE_PRICE_ID=
 NEXT_PUBLIC_APP_REDIRECT_URL=yourapp://checkout_redirect/success
 NEXT_PUBLIC_PADDLE_ENV=sandbox # or `production`
